Clean up the README_template file.
Since you're using str.format() in the README template for new books, you can
leverage how format works, which makes for cleaner code in GITenburg.py and
minor changes in the template. If you want the template to be more obvious,
replace 0 with 'book' and in the parameters to format use book=book.

Also I changed some styling in the README for GITenburg.