Set User-Agent string for GITenberg.py.
This will allow us to identify with the GitHub API to request a higher
rate-limit if we determine we need it. It's currently only been introduced
post 0.1b0's release. I'll release an 0.1b1 after GitHub ships their
Notifications API or at the end of October, which ever happens first.